fix compiler warnings

+++ floatmagic.h
@@ -32,9 +32,9 @@
 #ifndef HAVE_ISNAN
 #ifndef isnan
 # define isnan(x) \
-    (sizeof (x) == sizeof (long double) ? isnan_ld (x) \
-     : sizeof (x) == sizeof (double) ? isnan_d (x) \
-     : isnan_f (x))
+    (sizeof ((x)) == sizeof (long double) ? isnan_ld ((long double) (x)) \
+     : sizeof ((x)) == sizeof (double) ? isnan_d ((double) (x)) \
+     : isnan_f ((float) (x)))
 static inline int isnan_f  (float       x) { return x != x; }
 static inline int isnan_d  (double      x) { return x != x; }
 static inline int isnan_ld (long double x) { return x != x; }
@@ -44,9 +44,9 @@ static inline int isnan_ld (long double x) { return x != x; }
 #ifndef HAVE_ISINF
 #ifndef isinf
 # define isinf(x) \
-    (sizeof (x) == sizeof (long double) ? isinf_ld (x) \
-     : sizeof (x) == sizeof (double) ? isinf_d (x) \
-     : isinf_f (x))
+    (sizeof ((x)) == sizeof (long double) ? isinf_ld ((long double) (x)) \
+     : sizeof ((x)) == sizeof (double) ? isinf_d ((double) (x)) \
+     : isinf_f ((float) (x)))
 static inline int isinf_f  (float       x) { return isnan (x - x); }
 static inline int isinf_d  (double      x) { return isnan (x - x); }
 static inline int isinf_ld (long double x) { return isnan (x - x); }
